Introduce the status package
This package will contain the checkup's status that
will be reported to the user.
It embeds the `Status` struct, required by the kiagnose library.

The `Results` struct will contain the checkup's measurements.

This package is a dependency of the Launcher that will be
introduced in a follow-up commit.
Also adding pkg to the lint and test/unit makefile targets

Signed-off-by: Ram Lavi <ralavi@redhat.com>