checkup, tests: Check if VMI / Pod exists before deleting it
Currently, the existence of the VMI / Pod is not checked before
deleting the object.

Add a check if the VMI / Pod exists before deleting it, or
return an error in case it doesn't exist.

This uncovered a problem that was caused by the fact
the VMI / Pod namespace was not stored.

Store the VMI / Pod namespace upon their creation.

Signed-off-by: Orel Misan <omisan@redhat.com>