executer: Introduce trex-console object
This module wraps the pod execute command in order to run the
trex-console binary.
The trex-console is run with the -q (=quiet) flag in order to reduce
unneeded data on the stdout.
Moreover, performing basic cleaning of the stdout in order to filter out
the unneeded metadata.

The wrapper functions are using the commands and flags following the
trex-console documentation [0]

[0] https://trex-tgn.cisco.com/trex/doc/trex_console.html

Signed-off-by: Ram Lavi <ralavi@redhat.com>