trex-console: Get json format responses
In order to parse data retrieved from the trex-console,
adding the verbose option runCommand trex-console function.

This ability will expose the result json struct and will be turned
on in functions that return data. The json string is extracted from
the verbosed output, and then it is unmarshalled for easy handling.

Signed-off-by: Ram Lavi <ralavi@redhat.com>