checkup, vmi: Replace node selector with affinity
The VMI is added with a "kubevirt-dpdk-checkup/uid" label that will help
mark it for anti-affinity during the scheduling of the traffic-generator
pod.

If no node selector is chosen - the VMI is set to use the
"kubevirt-dpdk-checkup/uid" label for anti-affinity.

Using the "kubevirt-dpdk-checkup/uid" will allow running several
dpdk-checkup instances without affinity collision.

Signed-off-by: Ram Lavi <ralavi@redhat.com>