checkup, pod: Replace Node selector with affinity
The Pod is added with a "kubevirt-dpdk-checkup/uid" label that will help
mark
it for anti-affinity during the scheduling of the DPDK-VMI.

If no node selector is chosen - the traffic-generator pod is set to use
the
"kubevirt-dpdk-checkup/uid" label for anti-affinity.

Using the "kubevirt-dpdk-checkup/uid" will allow running several
dpdk-checkup instances without affinity collision.

Signed-off-by: Ram Lavi <ralavi@redhat.com>