executer, testpmd: Refactor extractSummaryStatistics to be generic
currently the extractSummaryStatistics is extracting only the summarized
section of the testpmd forwarding statistic output.
Enhancing the function to be more generic, and allow returning port
sections as well.
The method of Search via Index in the string is replaced with line
search in order to allow flexibility in the section start/end choosing.

Signed-off-by: Ram Lavi <ralavi@redhat.com>