traffic-gen: Align ports with DPDK VMI
currently the direction the traffic generator is sending the packets is
from port 0 using the west src MAC, but since the DPDK application is configured
to receive the east MAC on port 0, then the ports gets mixed.
This does not effect the nature of the checkup, but for the sake of
order, the checkup traffic diagram should be like this:
traffic-gen[port 0] => dpdk-app[port 0] => dpdk-app[port 1] traffic-gen[port 1].
Switching the MACs on the traffic generator streams templates.

Signed-off-by: Ram Lavi <ralavi@redhat.com>