config, checkup, executer: Change pps param
The trafficGeneratorPacketsPerSecondInMillions param is currently an
int and represents pps in millions.
In order to allow sending other rates (10m, 10k, 10) according to
trex-console docs [0], this parameter is changed to a string, and the
name is shortened to trafficGeneratorPacketsPerSecond.
A validation is added to make sure it is used properly.

[0] https://trex-tgn.cisco.com/trex/doc/trex_console.html, section
1.3.4. Multiplier

Signed-off-by: Ram Lavi <ralavi@redhat.com>