vm image builder: Upgrade to CentOS Stream9
Currently, we are building VM images using CentOS Stream 8
base image.
CentOS Stream 9 is the latest major version of CentOS Stream.
Bump the base image to CentOS Stream 9.

The `libguestfs-tools` package (which contains virt-builder)
had been renamed to `guestfs-tools`.

As a workaround for a bug[1], there is a need to run the
`kubevirt-dpdk-checkup-vm-image-builder` container as privileged.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2175778.

Signed-off-by: Orel Misan <omisan@redhat.com>