Update kubevirt.io/client-go to v1.0.0
Currently, we are using kubevirt.io/client-go@v0.58.0.

Update to v1.0.0 in order to:
1. Take advantage of context.Context parameter
   that was added to VMI methods.
2. Eliminate the following warning from the checkup's logs:
```
W0702 08:56:11.138067 1 warnings.go:70] kubevirt.io/v1alpha3
is now deprecated and will be removed in a future release.

There was a need to add `//nolint` to a deprecated
function call `Stream()`, that will be handled in a follow-up
PR.

Signed-off-by: Orel Misan <omisan@redhat.com>