checkup, vmi: Reduce Hugepage quantity demand to 4Gi
Currently the VMI is requesting 8 Hugepages of size 1Gi.
This is an unnecessary requirement.
In the VMI spec, removing the current memory request and adding instead
the Guest memory needed in order to ensure that the guest will be able
to allocate at least 1 Hugepage of size 1Gi at boot.
In the VMI cmvline boot command, update the Hugepage request to 4.

Note: Due to an issue in allocation with the HugeTLB, when requesting
4 hugapages allocated in the guest, only 1 is allcoated. However this is enough for our case.

Signed-off-by: Ram Lavi <ralavi@redhat.com>