Build system: Add traffic-gen VM
Add the files required to build a containerDisk image for the
traffic-gen VM.

These files are based of the VM-under-test files.

curl is used with the `--insecure` option, because without it
the following error message appears while building the qcow2 image
(using virt-builder)
```
curl: (60) SSL certificate problem: unable to get local issuer certificate
```

This should be addressed in the future.

tar is used with the `--strip-components=1` in order to strip
the archive's root directory which has the same name as the
TRex version.
The above will help make the checkup more agnostic to TRex version.

Signed-off-by: Orel Misan <omisan@redhat.com>