Checkup, test: Add param to clientStub.VMIName()
Currently, the checkup creates a single VMI object
that represents the VMI under test.

In preparation for the addition of a traffic generator VMI,
add a `prefix` parameter to the `clientStub.VMIName` method
so a future commit could query for the name of the traffic generator VMI.

Also rename all variables named `vmiName` to `vmiUnderTestName`.

Signed-off-by: Orel Misan <omisan@redhat.com>