executor, testpmd: Refactor PCI-Address parameter names
Now that both the traffic generator and "VMI under test" are VMIs, the PCI-Address
parameters are left ambiguous, because there is more than one VMI.

In order to keep the keep the logic fluency - these changes are done
both executer and testpmd packages in the same commit:
- Change the params' 'vmi' prefix to 'vmiUnderTest' on the executor
object params.
- Change the params' 'vmi' prefix to 'vmiUnderTest' on the testpmd
constuctor input params.
Inside the testpmd package context itself there is no ambiguity since it
is inside the vmi guest, so keeping the object params without change.

Signed-off-by: Ram Lavi <ralavi@redhat.com>