Checkup, Setup: Create traffic gen ConfigMap
In order to enable the checkup the ability to create a ConfigMap
containing the TRex[1] configuration to be attached to the
traffic gen VMI, add creation of an empty ConfigMap object to the
setup flow.

The above traffic gen ConfigMap is owned by the checkup, so it
will be garbage collected by K8s, in case of a failure.

[1] https://trex-tgn.cisco.com/

Signed-off-by: Orel Misan <omisan@redhat.com>