Checkup, vmi: Build VMIs using building blocks
Currently, the two VMIs are built using the same basic spec, but with
different parameters:
1. VMI name.
2. Affinity settings.
3. ContainerDisk image name.
4. NIC{1..2} MAC and PCI addresses.

The traffic gen VMI, needs to have additional configuration
in order to receive the TRex configuration using a ConfigMap object.

Separate the build of the two VMIs.
Use a base spec that is identical between the VMIs.

Signed-off-by: Orel Misan <omisan@redhat.com>