Checkup: Name all created entities with the same suffix
Currently, the two VMIs are given a different random suffix.
Also, the traffic gen ConfigMap object has no suffix,
which prevents running more than one instance of the checkup
in a single namespace.

Give the two VMIs and the traffic gen ConfigMap the same random
suffix, so it will be easier to identify then in a glance.

Signed-off-by: Orel Misan <omisan@redhat.com>