trex, executor: Run traffic towards VMI under test
Performing the run traffic command in the trex-console [0].
The traffic is defined in the stream python scripts uploaded to the VMI
on startup. flags used [1]:
-m: Multiplier, defines the traffic speed in pps.
-p: Port. defines the directionality of traffic.
-d: Duration.

Since the function is not blocking, sleep waiting until the test is
over. In next commits this sleep will be replaced with a function
monitoring the runtime params like dropRate.

[0] https://trex-tgn.cisco.com/trex/doc/trex_console.html section 1.2.3:
Transmitting traffic.
[1] https://trex-tgn.cisco.com/trex/doc/trex_console.html section 1.3:
Common Arguments

Signed-off-by: Ram Lavi <ralavi@redhat.com>