trex, executor: Replace test sleep with monitoring for dropRate
Use the "stats -g" trex-console command [0] in order to poll the
The values are printed to the log.

In order to parse trex commands outputs with statistics, introducing
another console-cmd function 'runTrexConsoleCmdWithJSONResponse' that
utilizes the verbose trex-console option, so the output will include the
output command statistics in JSON format as well. This format will be
used to acquire the desired parameters from the command query.

Unfortunately in order to clean the output from unprintable characters
that cause the JSON unmarshaling to fail - there was a need to exclude
one line from the gocritic linter.

[0] https://trex-tgn.cisco.com/trex/doc/trex_console.html section 1.2.4:
Getting the stats.

Signed-off-by: Ram Lavi <ralavi@redhat.com>