testpmd, console: Add vmiName field
Currently, the exported methods of `testpmd.Console`
are receiving a vmiName parameter.

Because this object is supposed to work against a single VMI,
a better approach would be to pass the vmiName parameter during
the object's construction.

Add a vmiName field to `testpmd.Console`.
Remove the vmiName parameter from all its exported methods.

Signed-off-by: Orel Misan <omisan@redhat.com>