console: Move SafeExpectBatchWithResponse to be a method of Expecter
The SafeExpectBatchWithResponse is a basic function of the console.
It allows the consumer to send console commands to the console using the
genExpect tool, while using all the internal params of the expecter
object. All of the above suggest it should be a method of the object itself.
Moving to be a method of the expecter object.

This change means that currently the expecter object is created
on each console method. This will be replaced in future commits
towards the testpmd object holding the expecter object itself.

Signed-off-by: Ram Lavi <ralavi@redhat.com>