console, executor: Move LoginToCentOS to a method of Expecter object
The LoginToCentOS is a basic function of the console. it starts of the
initial login and basic configuration logic, using all the internal
params of the expecter object. It does so using one spawned console
instead of opening multiple ones. All of the above suggest it should
be a method of the object itself.
Moving to be a method of the expecter object.

Signed-off-by: Ram Lavi <ralavi@redhat.com>