vm image builder: Upgrade to CentOS Stream 9
Currently, we are building VM images using CentOS Stream 8
base image.
CentOS Stream 9 is the latest major version of CentOS Stream.
Bump the base image to CentOS Stream 9.

The `libguestfs-tools` package (which contains virt-builder)
had been renamed to `guestfs-tools`.

It is important to note that the VM under test and the traffic gen VM
will still run on CentOS Stream 8.

Signed-off-by: Orel Misan <omisan@redhat.com>