E2E: Make NAD name mandatory
Currently, the `NETWORK_ATTACHMENT_DEFINITION_NAME`
environment variable is optional and has a default value
in case it is empty or does not exist.

Currently, there is no need for a default value for this
environment variable, because we are explicitly stating
its value while running the E2E test.

Make it mandatory.

Signed-off-by: Orel Misan <omisan@redhat.com>