Config: Add vmUnderTestAlwaysPullContainerDiskImage
Currently, the VM under test container disk image pull
policy is set to `PullAlways`.

This is necessary for testing, but negatively impacts
the checkup's setup time in production.

In order to set the default image pull policy
to `PullIfNotPresent` in a follow-up commit,
add the `vmUnderTestAlwaysPullContainerDiskImage` field
to allow the user to set the image pull policy to `PullAlways`
after the default policy will be changed to `PullIfNotPresent`.

Signed-off-by: Orel Misan <omisan@redhat.com>