E2E: Always pull container disk images
The container disk image pull policy of the two VMs was
changed by previous commits to `PullIfNotPresent`.

In order to make sure we are testing with the latest images,
set the following fields to "true"
 - `spec.param.trafficGenAlwaysPullContainerDiskImage`
 - `spec.param.vmUnderTestAlwaysPullContainerDiskImage`

the above configuration will make sure that the
container disk images will always be pulled during testing.

Signed-off-by: Orel Misan <omisan@redhat.com>