checkup: Aggregate teardown errors
Currently, the teardown logic is as follows:

- Remove the VMI under test
- Remove the traffic gen
- Remove the traffic gen ConfigMap
- Wait for VMI under test to be removed
- Wait for the traffic gen to be removed.
In case any of the above steps fails, the entire teardown fails.

Align this logic with the VM latency checkup, in which the errors are
accumulated.
Fix deletion unit test accordingly. The fix required adding a short
timeout context, because otherwise the wait for the VMI that failed to
be deleted hangs until timeout.

Signed-off-by: Ram Lavi <ralavi@redhat.com>