Checkup, setup: Add setup context
Currently, the setup is using the global context,
which has the timeout set by the user.

In case the checkup had created in VMIs, there is
a wait for both VMIs to be ready, that does not fail until
the global timeout.

If the user-supplied timeout is long, the checkup will wait
for a long time before failing.

Add a context with a ten minutes timeout for the setup phase,
derived from the global context.

Signed-off-by: Orel Misan <omisan@redhat.com>