testpmd: Set specific CPUs to run forwarding traffic
When configuring the cores that will run the traffic one needs to think
of the current topology.

Currently the vmi-under-test guest is using 8 CPUs: 0-7.
Hence the SMT siblings are set by kubevirt to be 0-1,2-3,4-5,6-7.
0-1 are set to be reserved to the OS on the grub linuxcmd [0].
2-3 should run the testpmd main thread (running non DPDK processes)
thus - 4,5,6,7 should be assigned to the forwarding CPUs (four in
total).

This commit is changing:
- The `-l` parameter to the `--lcores` in order to specifically set the
CPUs to the appropriate guest vCPUs [1]
- The numberOfCores used in `nb-cores` parameter to set the number of
forwarding CPUs to four.

[0]
https://github.com/kiagnose/kubevirt-dpdk-checkup/blob/e0deb7ead901ca189bba80ca139c6f39949b5d18/vms/vm-under-test/scripts/customize-vm#L25
[1] https://doc.dpdk.org/guides-16.04/testpmd_app_ug/run_app.html

Signed-off-by: Ram Lavi <ralavi@redhat.com>