trex: Set specific CPUs to run forwarding traffic
When configuring the cores that will run the traffic one needs to think
of the current topology.

Currently the traffic-gen-vmi guest is using 8 CPUs: 0-7.
Hence the SMT siblings are set by kubevirt to be 0-1,2-3,4-5,6-7.
0-1 are set to be reserved to the OS on the grub linuxcmd [0].
2-3 should run the trex master and latency threads (running non DPDK
processes)
thus - 4,5,6,7 should be assigned to the forwarding CPUs (four in
total).

This commit is changing these parameters in the trex server cmdline[1]:
* -c: the number is changed to four traffic forwarding CPUs.
Also, it is changing the trex_cfg.yaml file:
* update CPU mapping to use the appropriate vCPUs

[0]
https://github.com/kiagnose/kubevirt-dpdk-checkup/blob/e0deb7ead901ca189bba80ca139c6f39949b5d18/vms/traffic-gen/scripts/customize-vm#L27
[1] https://trex-tgn.cisco.com/trex/doc/trex_manual.html#cml-line

Signed-off-by: Ram Lavi <ralavi@redhat.com>