console/console: Disable console package verbose flag
In the checkup there are two independant vebose flags:
1. the console package verbose flag. hard coded to true.
2. the user-supplied verbose flag.

The console package verbose flag causes multiple prints to the logs per
each console batch command sent, and is redundant to the user-supplied
verbose flag prints that are tailor made for the checkup debugging,
making the log very messy.
Setting the console package verbose flag to be false.

Signed-off-by: Ram Lavi <ralavi@redhat.com>