trex/client: Unremove stdout lines
The trex/client is a wrapper for the trex-console binary on the guest.
The stdout of this binary catains many lines - that are currently
repressed because they are not used.
In the next commits, this trex-console output is going to be parsed, so
it makes sense to un-repress these lines to keep the original context of
the output.
Returning the trex-console output lines to avoid mutating the raw
output.

Signed-off-by: Ram Lavi <ralavi@redhat.com>