vms/vm-under-test: Move first boot to a separate service
The first-boot script running by the virt-builder command is run on a
service created by virt-builder. This service is guaranteed to run in
the final stages of the boot process [0].
This may create a race with the checkup, that is waiting on the
agentConnected condition [1] being added in order to run the checkup's
executor package. The race is happening since the guest-agent service
also runs during the final stages of the boot process.

In order to eliminate this race, removing the first-boot script in favor
of the new service, and moving the first boot script content
into a new service. This service is
- manually created on the customize-vm script.
- explicitly scheduled to run before the guest-agent service runs.
- running the content of the former first-boot script.
- running the tuned-adm command formerly set on the customize-vm script.
The reason is that this command needs to run on a running guest.
Moreover, since this command needs to only run on the first boot of the
image, the script is checking the existence of a marker file, then only
creating the file after the first run of the snippet.

[0] https://www.libguestfs.org/virt-builder.1.html
[1]
https://github.com/kiagnose/kubevirt-dpdk-checkup/blob/b9b6a472fe92583a0db46361289239e0e8d06284/pkg/internal/checkup/checkup.go#L237
[2] https://www.freedesktop.org/software/systemd/man/latest
systemd.unit.html#ConditionPathExists=

Signed-off-by: Ram Lavi <ralavi@redhat.com>