vms/customize-vm: Remove in-script reboot
The current approach in getting the vmi ready for the checkup is to set
the tuned-adm + reboot commands before the guest-agent runs. The reason
is that guest-agent ready is the signal that the checkup uses in order
to know if the VMI has successfully booted.
Rebooting before the guest-agent is ready is important to avoid
a race where the checkup continues to the test before the tuned-adm
kernel-args are set (requires reboot).
This approach is flawed. The scheduling mentioned above applies only to
the when the service starts, and does not ensure that it runs serially
after the guest-agent service. This means that the reboot can actually
be performed later in the systemd boot sequence - well after the
guest-agent service is started. Setting the --force flag in the reboot
command in order to hasten the reboot does not change this behavior.

This commit removes the reboot from the dpdk-checkup-boot script on both
vm images, in favor of a manual reboot done on the checkup itself, that
will be introduced in later commits.

Signed-off-by: Ram Lavi <ralavi@redhat.com>