vms/customize-vm: Remove dpdk-checkup-boot.service
Currently the image is set to add a service running the commands needed
for configuring the guest for DPDK. This service is set to schedule
before the guest-agent service, but in reality it takes longer for it to
finish running and by then the guest agent service is already up and
running.
Since the guest-agent service being ready is the criteria for the VMI
being booted and the checkup moving to the test execution phase, this
service finishing to run after guest-agent exposes the checkup for a
race where the checkup continues before the guest was properly
configured.
Hence, there is no point in running this service the way it does.
This commit removes the dpdk-checkup-boot.service in favor of adding it
to the cloud-init service in future commits. This has the benefit of
simplifying the image build process.

Signed-off-by: Ram Lavi <ralavi@redhat.com>