checkup: Add guest-agent-ping probe replace wait cond to ready state
Currently the checkup setup only waits for the VMI to finish "booting",
i.e. the guest agent service to be ready.
However this is not enough in order to ensure that the VMI has been
configured, a procedure currently done on the cloud-init service.
When the configuration is complete, the script configuring the guest in
the cloud-init service adds a marker file.
Changing the wait condition to "ready" is stronger than waiting for
"guest-agent", since the guest-agent ping needs guest agent in order to
set the VMI to ready.

This commit:
- introduces a new waiting mechanism, using guest-agent-ping probe [0]
to poll-wait the guest until the file is present, and only then sets the
VMI ready condition to true.
- adds a wait for the VMI ready condition to be true.
- removed the now unnecessary wait for guest-agent-ready condition.

[0]
https://kubevirt.io/user-guide/virtual_machines/liveness_and_readiness_probes/#defining-guest-agent-ping-probes

Signed-off-by: Ram Lavi <ralavi@redhat.com>
(cherry picked from commit ee2b43303666003ebc141ec2357803c9e286c7ad)