checkup: Replace reboot in boot script with Soft reboot
In order for the tuned-adm command running in the boot script to take
affect - there is a need for a reboot.
Currently this reboot is done on the boot script.
This commit is replacing this boot with an API soft reboot to the VMI,
after the VMI is declared ready.

Signed-off-by: Ram Lavi <ralavi@redhat.com>