checkup: Soft reboot VMI after VMI is ready
The guest-agent-ping wait-poll is waiting until the marker file,
indicating that the VMI has been properly configured - is set before
setting the VMI to condition ready.
This commit is adding a soft reboot to the VMI, after it is set to
ready.
The reboot is necessary in order for the tuned-adm command set on the
cloud-init service to take affect on the kernel args.

Signed-off-by: Ram Lavi <ralavi@redhat.com>