checkup: Add boot script to VMs' configmaps
Currently the VM guests are configured on a service added on the
customize-vm script created in the image building process.
There is no real need for the script to run on that context, and it is
easier to manage if it runs as part of the cloud-init context.
This commits:
- copies the script to a checkup function.
- adds the boot script to the traffic-gen and vmi-under-test VMs
- mounts the script with the correct permissions.

Signed-off-by: Ram Lavi <ralavi@redhat.com>
(cherry picked from commit 9dfcf78b9289771996c982c9220fc14d9f73a5b7)