makefile: push commit based tags when pushing vm-under-test image
Currently when a PR is merged then the vm-under-test disk-image is built
and pushed [0] to the kiagnose org on quay with the "latest" tag [1].
This commit adds another tag that is commit based (first 8 characters of
the sha commit [2]).
Adding it will allow fetch the images present on every future merged PR
on the repo.

[0]
https://github.com/kiagnose/kubevirt-dpdk-checkup/blob/ccc6f641cceaa088898571109031ec94be1040b3/.github/workflows/vm_containerdisk_publish.yaml

[1]
https://github.com/kiagnose/kubevirt-dpdk-checkup/blob/ccc6f641cceaa088898571109031ec94be1040b3/Makefile#L120C66-L120C93

[2]
https://github.com/kiagnose/kubevirt-dpdk-checkup/blob/ccc6f641cceaa088898571109031ec94be1040b3/Makefile#L5

Signed-off-by: Ram Lavi <ralavi@redhat.com>