Makefile: Remove spacial tag for each PR
Currently, for each PR pushed to main and for every tag created,
we push the checkup's image with a tag which is a substring
of the git commit hash.

This fills our quay registry with lots of unused tags, takes up
more storage space in quay, and adds noise when browsing the tags list.

Signed-off-by: Orel Misan <omisan@redhat.com>